+++ lib/mediawiki/query.rb
@@ -3,42 +3,41 @@ require_relative 'constants'
 
 module MediaWiki
   module Query
-    #TODO: Actually decide on a good way to deal with meta information queries.
-    # The metainformation could probably be handled in a much less verbose way.
-    # Perhaps we should get hashes instead?
+    # TODO: Actually decide on a good way to deal with meta information queries.
+    #   The metainformation could probably be handled in a much less verbose
+    #   way. Perhaps we should get hashes instead?
     module Meta
-
       # Returns an array of all the wiki's file repository names.
       # @return [Array] All wiki's file repository names.
       def get_filerepo_names
         params = {
           action: 'query',
           meta: 'filerepoinfo',
-          friprop: 'name',
-          format: 'json'
+          friprop: 'name'
         }
 
         result = post(params)
-        ret = Array.new
-        result["query"]["repos"].each do |repo|
-          ret.push(repo["name"])
+        ret = []
+        result['query']['repos'].each do |repo|
+          ret.push(repo['name'])
         end
-        return ret
+
+        ret
       end
 
       # Gets meta information for the currently logged in user.
       # @param prop [String] The uiprop to get.
-      # @return [Response/Boolean] Either a full, parsed response, or false if not logged in.
+      # @return [Response/Boolean] Either a full, parsed response, or false if
+      #   not logged in.
       def get_current_user_meta(prop)
         if @logged_in
           params = {
             action: 'query',
             meta: 'userinfo',
-            uiprop: prop,
-            format: 'json'
+            uiprop: prop
           }
 
-          return response = post(params)
+          return post(params)
         else
           return false
         end
@@ -46,8 +45,8 @@ module MediaWiki
     end
 
     module Properties
-
-      # Gets the wiki text for the given page. Returns nil if it for some reason cannot get the text, for example, if the page does not exist. Returns a string.
+      # Gets the wiki text for the given page. Returns nil if it for some
+      #   reason cannot get the text, for example, if the page does not exist.
       # @param title [String] The page title
       # @return [String/nil] String containing page contents, or nil
       def get_text(title)
@@ -55,19 +54,19 @@ module MediaWiki
           action: 'query',
           prop: 'revisions',
           rvprop: 'content',
-          format: 'json',
           titles: title
         }
 
         response = post(params)
-        response["query"]["pages"].each do |revid, data|
-          $revid = revid
+        revid = nil
+        response['query']['pages'].each do |r, _|
+          revid = r
         end
 
-        if response["query"]["pages"][$revid]["missing"] == ""
+        if response['query']['pages'][revid]['missing'] == ''
           return nil
         else
-          return response["query"]["pages"][$revid]["revisions"][0]["*"]
+          return response['query']['pages'][revid]['revisions'][0]['*']
         end
       end
 
@@ -79,13 +78,12 @@ module MediaWiki
           action: 'query',
           prop: 'revisions',
           rvprop: 'content',
-          format: 'json',
           titles: title
         }
 
         response = post(params)
-        response["query"]["pages"].each do |revid, data|
-          if revid != "-1"
+        response['query']['pages'].each do |revid, _|
+          if revid != '-1'
             return revid.to_i
           else
             return nil
@@ -93,28 +91,31 @@ module MediaWiki
         end
       end
 
-      # Gets the edit token for the given page. This method should rarely be used by normal users.
-      # @param page_name [String] The page title that you are going to be editing.
-      # @return [String] The edit token. If the butt isn't logged in, it returns with '+\\'.
+      # Gets the edit token for the given page. This method should rarely be
+      #   used by normal users.
+      # @param page_name [String] The page title that you are going to be
+      #   editing.
+      # @return [String] The edit token. If the butt isn't logged in, it returns
+      #   with '+\\'.
       def get_edit_token(page_name)
         if @logged_in == true
           params = {
             action: 'query',
             prop: 'info',
             intoken: 'edit',
-            format: 'json',
             titles: page_name
           }
 
           response = post(params)
-          response["query"]["pages"].each do |revid, data|
-            $revid = revid
+          revid = nil
+          response['query']['pages'].each do |r, _|
+            revid = r
           end
 
           # URL encoding is not needed for some reason.
-          return response["query"]["pages"][$revid]["edittoken"]
+          return response['query']['pages'][revid]['edittoken']
         else
-          return "+\\"
+          return '+\\'
         end
       end
     end
@@ -124,13 +125,14 @@ module MediaWiki
 
       # Gets an array of backlinks to a given title.
       # @param title [String] The page to get the backlinks of.
-      # @param limit [Int] The maximum number of pages to get. Defaults to 500, and cannot be greater than that unless the user is a bot. If the user is a bot, the limit cannot be greater than 5000.
+      # @param limit [Int] The maximum number of pages to get. Defaults to 500,
+      #   and cannot be greater than that unless the user is a bot. If the user
+      #   is a bot, the limit cannot be greater than 5000.
       # @return [Array] All backlinks until the limit
       def what_links_here(title, limit = 500)
         params = {
           action: 'query',
-          bltitle: title,
-          format: 'json'
+          bltitle: title
         }
 
         if limit > 500
@@ -147,24 +149,28 @@ module MediaWiki
           params[:bllimit] = limit
         end
 
-        ret = Array.new
+        ret = []
         response = post(params)
-        response["query"]["backlinks"].each do |bl|
-          ret.push(bl["title"])
+        response['query']['backlinks'].each do |bl|
+          ret.push(bl['title'])
         end
-        return ret
+
+        ret
       end
 
       # Returns an array of all page titles that belong to a given category.
-      # @param category [String] The category title. It can include "Category:", or not, it doesn't really matter because we will add it if it is missing.
-      # @param limit [Int] The maximum number of members to get. Defaults to 500, and cannot be greater than that unless the user is a bot. If the user is a bot, the limit cannot be greater than 5000.
+      # @param category [String] The category title. It can include "Category:",
+      #   or not, it doesn't really matter because we will add it if it is
+      #   missing.
+      # @param limit [Int] The maximum number of members to get. Defaults to
+      #   500, and cannot be greater than that unless the user is a bot.
+      #   If the user is a bot, the limit cannot be greater than 5000.
       # @return [Array] All category members until the limit
       def get_category_members(category, limit = 500)
         params = {
           action: 'query',
           list: 'categorymembers',
-          cmprop: 'title',
-          format: 'json'
+          cmprop: 'title'
         }
 
         if category =~ /[Cc]ategory\:/
@@ -187,26 +193,28 @@ module MediaWiki
           params[:cmlimit] = limit
         end
 
-        ret = Array.new
+        ret = []
         response = post(params)
-        response["query"]["categorymembers"].each do |cm|
-          ret.push(cm["title"])
+        response['query']['categorymembers'].each do |cm|
+          ret.push(cm['title'])
         end
         ret
       end
 
       # Returns an array of random pages titles.
-      # @param number_of_pages [Int] The number of articles to get. Defaults to 1. Cannot be greater than 10 for normal users, or 20 for bots.
-      # @param namespace [Int] The namespace ID. Defaults to '0' (the main namespace). Set to nil for all namespaces.
+      # @param number_of_pages [Int] The number of articles to get.
+      #   Defaults to 1. Cannot be greater than 10 for normal users,
+      #   or 20 for bots.
+      # @param namespace [Int] The namespace ID. Defaults to
+      #   0 (the main namespace).
       # @return [Array] All members
       def get_random_pages(number_of_pages = 1, namespace = 0)
         params = {
           action: 'query',
-          list: 'random',
-          format: 'json'
+          list: 'random'
         }
 
-        if $namespaces.has_value?(namespace)
+        if $namespaces.value?(namespace)
           params[:rnnamespace] = namespace
         else
           params[:rnnamespace] = 0
@@ -226,18 +234,22 @@ module MediaWiki
           params[:rnlimit] = number_of_pages
         end
 
-        ret = Array.new
+        ret = []
         responce = post(params)
-        responce["query"]["random"].each do |a|
-          ret.push(a["title"])
+        responce['query']['random'].each do |a|
+          ret.push(a['title'])
         end
-        return ret
+
+        ret
       end
 
-      # Gets user information. This method should rarely be used by normal users.
+      # Gets user information. This method should rarely be used by
+      #   normal users.
       # @param prop [String] The usprop parameter.
-      # @param username [String] The username to get info for. Optional. Defaults to the currently logged in user if ommitted.
-      # @return [String/Nil] Parsed full response if successful, nil if the username is nil and the Butt is not logged in.
+      # @param username [String] The username to get info for. Optional.
+      #   Defaults to the currently logged in user if ommitted.
+      # @return [String/Nil] Parsed full response if successful, nil if
+      #   the username is nil and the Butt is not logged in.
       def get_userlists(prop, username = nil)
         if username.nil?
           if @logged_in
@@ -250,25 +262,26 @@ module MediaWiki
             action: 'query',
             list: 'users',
             usprop: prop,
-            ususers: username,
-            format: 'json'
+            ususers: username
           }
 
           response = post(params)
         end
 
-        return response
+        response
       end
 
       # Gets an array of all the user's groups.
-      # @param username [String] The username to get groups of. Optional. Defaults to the currently logged in user.
-      # @return [Array/Boolean] All of the user's groups, or false if username is nil and Butt is not logged in.
+      # @param username [String] The username to get groups of. Optional.
+      #   Defaults to the currently logged in user.
+      # @return [Array/Boolean] All of the user's groups, or false if username
+      #   is nil and Butt is not logged in.
       def get_usergroups(username = nil)
-        ret = Array.new
+        ret = []
         if username.nil?
           if @logged_in
             info = get_userlists('groups')
-            info["query"]["userinfo"]["groups"].each do |i|
+            info['query']['userinfo']['groups'].each do |i|
               ret.push(i)
             end
           else
@@ -276,25 +289,27 @@ module MediaWiki
           end
         else
           info = get_userlists('groups', username)
-          info["query"]["users"].each do |i|
-            i["groups"].each do |g|
+          info['query']['users'].each do |i|
+            i['groups'].each do |g|
               ret.push(g)
             end
           end
         end
 
-        return ret
+        ret
       end
 
       # Gets the user rights for the user.
-      # @param username [String] The user to get the rights for. Optional. Defaults to the currently logged in user.
-      # @return [Array/Boolean] All of the user's groups, or false if username is nil and Butt is not logged in.
+      # @param username [String] The user to get the rights for. Optional.
+      #   Defaults to the currently logged in user.
+      # @return [Array/Boolean] All of the user's groups, or false if username
+      #   is nil and Butt is not logged in.
       def get_userrights(username = nil)
-        ret = Array.new
+        ret = []
         if username.nil?
           if @logged_in
             info = get_userlists('rights')
-            info["query"]["userinfo"]["rights"].each do |i|
+            info['query']['userinfo']['rights'].each do |i|
               ret.push(i)
             end
           else
@@ -302,33 +317,38 @@ module MediaWiki
           end
         else
           info = get_userlists('rights', username)
-          info["query"]["users"].each do |i|
-            i["rights"].each do |g|
+          info['query']['users'].each do |i|
+            i['rights'].each do |g|
               ret.push(g)
             end
           end
         end
 
-        return ret
+        ret
       end
 
       # Gets contribution count for the user.
-      # @param username [String] The username to get the contribution count of. Optional. Defaults to the currently logged in user.
-      # @param autoparse [Boolean] Whether to automatically format the string with commas. Defaults to true.
-      # @return [Boolean/Int/String] False if username is nil and Butt is not logged in. An integer value of the contribution count if autoparse is false. A formatted string version of the contribution count if autoparse is true.
+      # @param username [String] The username to get the contribution count of.
+      #   Optional. Defaults to the currently logged in user.
+      # @param autoparse [Boolean] Whether to automatically format the string
+      #   with commas using string-utility. Defaults to true.
+      # @return [Boolean/Int/String] False if username is nil and Butt is not
+      #   logged in. An integer value of the contribution count if autoparse is
+      #   false. A formatted string version of the contribution count if
+      #   autoparse is true.
       def get_contrib_count(username = nil, autoparse = true)
         count = nil
         if username.nil?
           if @logged_in
             info = get_userlists('editcount')
-            count = info["query"]["userinfo"]["editcount"]
+            count = info['query']['userinfo']['editcount']
           else
             return false
           end
         else
           info = get_userlists('editcount', username)
-          info["query"]["users"].each do |i|
-            count = i["editcount"]
+          info['query']['users'].each do |i|
+            count = i['editcount']
           end
         end
 
@@ -336,11 +356,13 @@ module MediaWiki
           countstring = count.to_s.separate
           return countstring
         end
-        return count
+
+        count
       end
 
       # Gets when the user registered.
-      # @param username [String] The username to get the registration date and time of. Optional. Defaults to the currently logged in user.
+      # @param username [String] The username to get the registration date and
+      #   time of. Optional. Defaults to the currently logged in user.
       # @return [DateTime] The registration date and time as a DateTime object.
       def get_registration_time(username = nil)
         time = nil
@@ -348,48 +370,53 @@ module MediaWiki
         if username.nil?
           if @logged_in
             info = get_userlists('registrationdate')
-            time = info["query"]["userinfo"]["registrationdate"]
+            time = info['query']['userinfo']['registrationdate']
           else
             return false
           end
         else
           info = get_userlists('registration', username)
-          info["query"]["users"].each do |i|
-            time = i["registration"]
+          info['query']['users'].each do |i|
+            time = i['registration']
           end
         end
 
-        # %Y: Year including century, %m: Month num, %d day of month, %T Time as H:M:S
-        timeformat = "%Y-%m-%dT%T"
+        # %Y: Year including century
+        # %m: Month num
+        # %d: day of month
+        # %T: Time as HH:MM:SS
+        timeformat = '%Y-%m-%dT%T'
         time = DateTime.strptime(time, timeformat)
+
+        time
       end
 
       # Gets the gender for the provded user.
       # @param username [String] The user.
-      # @return [String] The gender. "male", "female", or "unknown".
+      # @return [String] The gender. 'male', 'female', or 'unknown'.
       def get_user_gender(username)
         gender = nil
         info = get_userlists('gender', username)
-        info["query"]["users"].each do |i|
-          gender = i["gender"]
+        info['query']['users'].each do |i|
+          gender = i['gender']
         end
 
-        return gender
+        gender
       end
 
       # Gets the amount of results for the search value.
       # @param search_value [String] The thing to search for.
-      # @param namespace [Int] The namespace to search in. Defaults to the main namespace.
+      # @param namespace [Int] The namespace to search in.
+      #   Defaults to 0 (the main namespace).
       # @return [Int] The number of pages that matched the search.
       def get_search_result_amount(search_value, namespace = 0)
         params = {
           action: 'query',
           list: 'search',
-          srsearch: search_value,
-          format: 'json'
+          srsearch: search_value
         }
 
-        if $namespaces.has_value?(namespace)
+        if $namespaces.value?(namespace)
           params[:srnamespace] = namespace
         else
           params[:srnamespace] = 0
@@ -397,22 +424,22 @@ module MediaWiki
 
         response = post(params)
 
-        return response["query"]["searchinfo"]["totalhits"]
+        response['query']['searchinfo']['totalhits']
       end
 
       # Gets an array containing page titles that matched the search.
       # @param search_value [String] The thing to search for.
-      # @param namespace [Int] The namespace to search in. Defaults to the main namespace.
+      # @param namespace [Int] The namespace to search in.
+      #   Defaults to 0 (the main namespace).
       # @return [Array] The page titles that matched the search.
       def get_search_results(search_value, namespace = 0)
         params = {
           action: 'query',
           list: 'search',
-          srsearch: search_value,
-          format: 'json'
+          srsearch: search_value
         }
 
-        if $namespaces.has_value?(namespace)
+        if $namespaces.value?(namespace)
           params[:srnamespace] = namespace
         else
           params[:srnamespace] = 0
@@ -420,12 +447,12 @@ module MediaWiki
 
         response = post(params)
 
-        ret = Array.new
-        response["query"]["search"].each do |search|
-          ret.push(search["title"])
+        ret = []
+        response['query']['search'].each do |search|
+          ret.push(search['title'])
         end
 
-        return ret
+        ret
       end
     end
   end